backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fonttools==4.61.1
h11==0.16.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.2.3
opencv-python-headless==4.12.0.88
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
